Write to SD card at specific interval
A new file was not being created at midnight if the device was
unplugged and plugged back in due to the time not being
syncronized to hit midnight at exactly 00:00:00.

I added a 5 second interval within the a loop that runs every
second. Now the file will be created at midnight since the loop
runs every second.