Several changes made
The csv file had too many sensor readings in it taken every
five seconds.

The sensor readings, date, and time will be written to the SD card
every ten minutes instead of every five seconds.

Added features:

Error messages will be printed to the character LCD if
there is a problem with the SD card.

Added a variable that keeps a count of the line numbers in the
first column of the csv file.

Added a header to the csv file so that people will know what each
column represents.